Remove deprecated `saveTrackedSlice` function and add Jet-generated models and tables for persistence layer
- Deleted `saveTrackedSlice` function as it is no longer in use.
- Integrated Go Jet-generated models and tables for entities like currency rates, families, family members, labels, providers, subscriptions, and more.
- These changes improve readability, maintainability, and alignment with database schema.

+++ backend/cmd/api/endpoints/subscription_get_all.go
@@ -31,8 +31,8 @@ type subscriptionGetAllQueryParams struct {
 	Users        []uuid.UUID `form:"users"`
 	WithInactive bool        `form:"with_inactive"`
 	Providers    []uuid.UUID `form:"providers"`
-	Limit        int32       `form:"limit"`
-	Offset       int32       `form:"offset"`
+	Limit        int64       `form:"limit"`
+	Offset       int64       `form:"offset"`
 }
 
 // Handle godoc
